Add SkipLink component (#84)
* feat: Add SkipLink component for improved accessibility

The SkipLink component is added to improve accessibility by providing a way for users to skip to the main content of the page. This component includes a link that is hidden visually but can be focused by screen readers. It also includes localization support with a default label. The component is implemented using Vue 3 composition API and is compatible with both Nuxt.js and standalone Vue applications.

* fix tests

* fix doc link

* lint

* fix index exports

* fix

---------

Co-authored-by: David FYON <davidfyon@gmail.com>